refactor: 리뷰 반영 1. Optional isPrsent() -> orElse() 2. @InjectMock 사용 x
