+++ src/main/java/com/sproutt/eussyaeussyaapi/api/dto/JoinDTO.java
@@ -1,6 +1,5 @@
 package com.sproutt.eussyaeussyaapi.api.dto;
 
-import com.sproutt.eussyaeussyaapi.domain.Member;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,14 +19,5 @@ public class JoinDTO {
         this.memberId = memberId;
         this.name = name;
         this.password = password;
-    }
-
-    public Member toEntity() {
-
-        return Member.builder()
-                .memberId(this.memberId)
-                .name(this.name)
-                .password(this.password)
-                .build();
     }
 }
