+++ src/test/java/com/sproutt/eussyaeussyaapi/application/MemberServiceTest.java
@@ -1,18 +1,18 @@
 package com.sproutt.eussyaeussyaapi.application;
 
-import com.sproutt.eussyaeussyaapi.domain.exceptions.DuplicatedMemberIdException;
+import com.sproutt.eussyaeussyaapi.api.dto.JoinDTO;
 import com.sproutt.eussyaeussyaapi.domain.Member;
 import com.sproutt.eussyaeussyaapi.domain.MemberRepository;
-import com.sproutt.eussyaeussyaapi.api.dto.JoinDTO;
+import com.sproutt.eussyaeussyaapi.domain.exceptions.DuplicatedMemberIdException;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -21,11 +21,13 @@ public class MemberServiceTest {
     private static final String NAME = "test";
     private static final String PASSWORD = "1111";
 
-    @InjectMocks
-    private MemberServiceImpl memberService;
+    private MemberRepository memberRepository = mock(MemberRepository.class);
+    private MemberService memberService;
 
-    @Mock
-    private MemberRepository memberRepository;
+    @BeforeEach
+    void setUp() {
+        memberService = new MemberServiceImpl(memberRepository);
+    }
 
     @Test
     public void createMember_with_exist_memberId() {
@@ -35,9 +37,17 @@ public class MemberServiceTest {
                 .password(PASSWORD)
                 .build();
 
-        Member member = joinDTO.toEntity();
+        Member member = defaultMember();
         when(memberRepository.findByMemberId(MEMBER_ID)).thenReturn(Optional.of(member));
 
         assertThrows(DuplicatedMemberIdException.class, () -> memberService.join(joinDTO));
+    }
+
+    private Member defaultMember() {
+        return Member.builder()
+                .memberId(MEMBER_ID)
+                .password(PASSWORD)
+                .name(NAME)
+                .build();
     }
 }
